Refactor marketdata schema boundaries
