+++ tests/marketdata/test_marketdata_schemas.py
@@ -5,18 +5,22 @@ from datetime import UTC, datetime
 import pandas as pd
 import pytest
 
+from option_pricing.marketdata.contracts import RunMetadata
+from option_pricing.marketdata.manifests import (
+    MODEL_VALIDATION_MANIFEST_REQUIRED_FIELDS,
+    validate_model_validation_manifest,
+)
 from option_pricing.marketdata.schemas import (
     CLEANED_QUOTES_COLUMNS,
     MODEL_VALIDATION_BUNDLE_VERSION,
-    MODEL_VALIDATION_MANIFEST_REQUIRED_FIELDS,
     DatasetName,
-    RunMetadata,
+)
+from option_pricing.marketdata.validation import (
     coerce_frame,
     dataset_columns,
     dataset_dtypes,
     validate_columns,
     validate_dtypes,
-    validate_model_validation_manifest,
 )
